Fix button (menu/direct) modes
ButtonMode needed to be SettingsObserver to see changes to LP_R/LP_B;
  tidy SetupBoxes; redraw to see effect immediately on setting change

Fix scanning (used wrong type for Time)

Scan/select if click in top/bottom half of screen in menu mode
  (or just select anywhere, if auto-scanning).

+++ Src/DasherCore/DasherButtons.cpp
@@ -36,7 +36,7 @@ void CDasherButtons::Activate() {
   // which depends on it, tho...
   if (!m_pBoxes) SetupBoxes();
 
-  m_iScanTime = std::numeric_limits<int>::min();
+  m_iScanTime = std::numeric_limits<unsigned long>::min();
 }
 
 void CDasherButtons::KeyDown(int iTime, int iId, CDasherView *pView, CDasherInput *pInput, CDasherModel *pModel) {
